+++ include/bot.h
@@ -60,9 +60,9 @@ typedef struct BotState {
   size_t move_scores_cap;
   int* move_scores;
   size_t fill_grid1_cap;
-  int* fill_grid1;
+  short* fill_grid1;
   size_t fill_grid2_cap;
-  int* fill_grid2;
+  short* fill_grid2;
   size_t fill_stack_cap;
   FillSpan* fill_stack;
 
@@ -84,8 +84,8 @@ BotMove* bot_generate_all_moves_list(
 int* bot_rate_moves_list(BotState* self, int moves_count, BotMove* moves_list);
 int bot_rate_move(BotState* self, BotMove move);
 bool bot_quick_junction_check(BotState* self, Coords coords);
-int* bot_flood_fill_reset_grid(BotState* self, int** fill_grid, size_t* fill_grid_cap);
-int bot_flood_fill_count_fish(BotState* self, int* grid, Coords start, int marker_value);
+short* bot_flood_fill_reset_grid(BotState* self, short** fill_grid, size_t* fill_grid_cap);
+int bot_flood_fill_count_fish(BotState* self, short* grid, Coords start, short marker_value);
 
 void flood_fill(
   int x,
